+++ CHANGELOG.md
@@ -45,6 +45,7 @@
   - Enable verbose preference to be passed through to `Test-DscParameterState`.
   - Changed `Test-TargetResource` so that `StartTime` is only compared for
     trigger types `Daily`,`Weekly` or `Once`.
+- Fix minor style issues in statement case.
 
 ## 6.5.0.0